InputとCaptureをcomponentにした
